requirements.txt
flask==2.2.5
flask-sqlalchemy==3.1.1
gunicorn==23.0.0
flask-login==0.6.3
flask-babel==2.0.0  # Version spécifique qui fonctionne avec Flask 2.2.5
flask-restful
flask-cors==5.0.1
flask-migrate==4.1.0
alembic==1.15.2
networkx==3.3
mako==1.3.10

python-dotenv
